Bug 943338 - Merge test_cards_view_launch tests
