+++ tests/python/gaia-ui-tests/gaiatest/tests/functional/cards_view/manifest.ini
@@ -2,10 +2,8 @@
 # Test requires B2G
 b2g = true
 
-[test_cards_view_launch_app.py]
-
 [test_cards_view_kill_apps_with_two_apps.py]
 
-[test_cards_view_with_three_apps.py]
-#  Bug 915138 - Cards view can have a maximum of 3 cards
-fail-if = device == "msm7627a"
+[test_cards_view_with_two_apps.py]
+#  Bug #946130 - Flick not working on select time - hour 24 picker
+expected = fail
